🐛 Fix bug when running with build1495 or newer
During development, I was still on build1494 which did not include Pebble, so I
included Pebble in frimgur. When creating an instance of PebbleEngine, a
default template loader using the class loader that Pebble was loaded with is
created which can access templates in the frimgur JAR if Pebble was also loaded
from the frimgur JAR; this was the case in build1494 but not anymore starting
from build1495.