🚧 Remove all notion of an image having a type while it’s not inserted
“It’s a surprise tool that’s going to help us later!”

In the Glorious Future™, the type of the image will be selectable in the
web interface and will be finalized by the insert. Before the insert,
the image will be stored in the raw data that it has been uploaded in
(which for me/Firefox always seems to be PNG, a prudent choice).